move core behaviors to the core's file
